common: move LoadLitFile to litfile.hh
