+++ include/common/bsputils.hh
@@ -176,6 +176,5 @@ public:
 
 qvec3b LM_Sample(const mbsp_t *bsp, const std::vector<uint8_t> *lit, const faceextents_t &faceextents,
     int byte_offset_of_face, qvec2i coord);
-std::vector<uint8_t> LoadLitFile(const fs::path &path);
 
 std::map<int, std::vector<int>> ClusterToLeafnumsMap(const mbsp_t *bsp);
